Create README.md
Background: 
The client, an online grocery store owner, had a pool of 5000 customers who made purchases in the past two years. He sought to profile these customers to create a targeted marketing strategy. The client believed that the best way to accomplish this was to utilize a machine learning model known as Customer Lifetime Value. He approached us to build the model, but he needed extensive guidance and was uncertain of the specific data points required for the model.

Problem: 
Given the client's lack of experience in this area, extensive back-and-forth communication was needed to acquire the necessary data.

Solution: 
We implemented an RFM (Recency, Frequency, Monetary) analysis to evaluate the days since each customer's last purchase, the number of purchases made, and the average monetary value of all purchases combined. This allowed the client to create detailed customer profiles based on the three parameters. We then calculated the probability of each customer making a future purchase by performing hyperparameter tuning. Finally, we developed a model to predict the total value of goods that each customer was likely to purchase in their lifetime (CLTV), providing the client with a method of distinguishing between high-value and low-value customers.
The client was thrilled to learn that we had developed three separate methods to profile his customers, rather than just the one he initially requested. He has since become a repeat customer, requesting the same algorithm for his other businesses and for his friend's businesses.